Move behavior module
