+++ lib/memcached/client.rb
@@ -1,52 +1,7 @@
 require 'memcached/marshal_codec'
+require 'memcached/behaviors'
 
 module Memcached
-
-  module Behaviors
-    CONVERSION_FACTORS = {
-      MEMCACHED_BEHAVIOR_RCV_TIMEOUT => 1_000_000,
-      MEMCACHED_BEHAVIOR_SND_TIMEOUT => 1_000_000,
-      MEMCACHED_BEHAVIOR_POLL_TIMEOUT => 1_000,
-      MEMCACHED_BEHAVIOR_CONNECT_TIMEOUT => 1_000
-    }
-
-    BASE_VALUES = {
-      MEMCACHED_BEHAVIOR_HASH => "MEMCACHED_HASH_%s",
-      MEMCACHED_BEHAVIOR_DISTRIBUTION => "MEMCACHED_DISTRIBUTION_%s",
-    }
-
-    def lookup_value(behavior, value)
-      base_value = BASE_VALUES[behavior]
-      raise ArgumentError unless base_value
-      value = base_value % value.to_s.upcase
-      int_value = Behaviors.const_get(value)
-      int_value * (CONVERSION_FACTORS[behavior] || 1)
-    end
-
-    def set_behaviors(hash)
-      hash.each do |key, value|
-        behavior_string = "MEMCACHED_BEHAVIOR_#{key.to_s.upcase}"
-        begin
-          behavior = Behaviors.const_get(behavior_string)
-        rescue NameError
-          raise ArgumentError, "No behavior #{behavior_string.inspect}"
-        end
-
-        if value.is_a? Symbol
-          begin
-            value = lookup_value(behavior, value)
-          rescue NameError, ArgumentError
-            msg =  "Invalid behavior value #{value.inspect} for #{behavior_string.inspect}"
-            raise(ArgumentError, msg)
-          end
-        end
-
-        set_behavior(behavior, value)
-      end
-
-    end
-  end
-
   class Connection
     include Behaviors
   end
